revert rememberComponentsActionHandler
